requirements.txt
# Core Dependencies
openai>=1.0.0
python-dotenv>=1.0.0
pyyaml>=6.0

# Vector Database & Embeddings
chromadb>=0.4.0
sentence-transformers>=2.2.0

# Document Processing
pypdf>=3.17.0
python-docx>=1.0.0
beautifulsoup4>=4.12.0
markdown>=3.5.0

# Retrieval & Search
rank-bm25>=0.2.2
nltk>=3.8.0

# Data & Analytics
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Dashboard & Visualization
streamlit>=1.28.0
plotly>=5.17.0
altair>=5.1.0

# Utilities
tqdm>=4.66.0
requests>=2.31.0
aiohttp>=3.9.0
